bonus

+++ libft/ft_putendl_fd.c
@@ -6,15 +6,15 @@
 /*   By: ervsahin <ervsahin@student.42kocaeli.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/10 07:19:17 by ervsahin          #+#    #+#             */
-/*   Updated: 2025/06/14 09:18:49 by ervsahin         ###   ########.fr       */
+/*   Updated: 2025/06/21 18:13:16 by ervsahin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <unistd.h>
 
 void	ft_putendl_fd(char *s, int fd)
 {
-	size_t	i;
+	int	i;
 
 	if (!s)
 		return ;
